Move contents of pkg into internal
