+++ internal/generate/gocode/gotemplate/gotemplate_test.go
