+++ internal/generate/gocode/gotemplate/http.go
@@ -1,4 +1,4 @@
-package http
+package gotemplate
 
 import (
 	"bytes"
