+++ internal/generate/gocode/gotemplate/tmpl/client.go.tmpl
