+++ internal/generate/gocode/gotemplate/tmpl/server.go.tmpl
