+++ internal/generate/gocode/gotemplate/tmpl/type.go.tmpl
