+++ internal/generate/gocode/testdata/client.go.golden
