+++ internal/generate/gocode/testdata/server.go.golden
