+++ internal/generate/gocode/testdata/service.cue
