+++ internal/generate/gocode/testdata/types.go.golden
